The v4l2_input driver can now get the capabilities of the video card, enumerate the inputs, and select the requested input.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@433 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_v4l2_input.h
@@ -24,7 +24,7 @@ typedef struct {
 	long input_number;
 
 	int fd;		/* File descriptor for the video device. */
-	int num_inputs;
+	long num_inputs;
 } MX_V4L2_INPUT;
 
 #define MXD_V4L2_INPUT_STANDARD_FIELDS \
